Envio autmático de Emails - Validações de estoque e controle de cliente

+++ backend/xmarket-backend-main/src/main/java/br/com/seguranca/services/ProdutoServico.java
@@ -9,11 +9,11 @@ import java.util.List;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
 import br.com.seguranca.dto.ProdutoDTO;
 import br.com.seguranca.model.Produto;
 import br.com.seguranca.repositories.ProdutoRepositorio;
-import org.springframework.web.multipart.MultipartFile;
 
 @Service
 public class ProdutoServico {
@@ -21,44 +21,32 @@ public class ProdutoServico {
 	@Autowired
 	private ProdutoRepositorio produtoRepository;
 
-
-
-	public Produto inserirProduto(ProdutoDTO produtoDTO) {
-		Produto produto = new Produto();
-		BeanUtils.copyProperties(produtoDTO, produto);
-
-
-
-		return produtoRepository.save(produto);
-	}
-	
-	public Integer buscarQuantidade(Long id){
-		return produtoRepository.buscarQuantidade(id);
-
-	}
-	
-	public List<ProdutoDTO> listarProdutos (){
-		List <Produto> produtos = new ArrayList<>();
+	public List<ProdutoDTO> listarProdutos() {
+		List<Produto> produtos = new ArrayList<>();
 		List<ProdutoDTO> produtoDTOS = new ArrayList<>();
 		produtos = produtoRepository.findAll();
-		for(Produto p : produtos){
+		for (Produto p : produtos) {
 			produtoDTOS.add(p.toProdutoDTO());
 		}
 		return produtoDTOS;
 	}
-	
-	public Produto inserirSomenteProduto(ProdutoDTO produtoDTO){
+
+	public boolean inserirSomenteProduto(ProdutoDTO produtoDTO) {
 		Produto produto = new Produto();
 		BeanUtils.copyProperties(produtoDTO, produto);
-
-		return produtoRepository.save(produto);
+		if (quantidadeValida(produto)) {
+			produtoRepository.save(produto);
+			return true;
+		}
+		return false;
 	}
-	
+
 	public String inserirSomenteImagem(MultipartFile arquivoImagem) {
 		Path path = Paths.get("");
 		try {
-			//"C:\\Users\\00787701\\Desktop\\github\\xmarket-1\\frontend\\xmarket-front\\assets\\produtos"
-			String diretorioImagens = "..\\..\\imagens-produtos\\"; //frontend/xmarket-front/assets/produtos/  //   ..\\..\\frontend\\xmarket-front\\assets\\produtos\\
+			// "C:\\Users\\00787701\\Desktop\\github\\xmarket-1\\frontend\\xmarket-front\\assets\\produtos"
+			String diretorioImagens = "..\\..\\imagens-produtos\\"; // frontend/xmarket-front/assets/produtos/ //
+																	// ..\\..\\frontend\\xmarket-front\\assets\\produtos\\
 			byte[] bytes = arquivoImagem.getBytes();
 			path = Paths.get(diretorioImagens + arquivoImagem.getOriginalFilename());
 			Files.write(path, bytes);
@@ -67,38 +55,43 @@ public class ProdutoServico {
 			e.printStackTrace();
 		}
 
-		return(path.toString());
+		return (path.toString());
 	}
 
 	public boolean excluirProduto(Long id) {
-		produtoRepository.deleteById(id);
-		return true;
-	}
-	
-	
-	public boolean editarProduto(Produto produto) {
-		if(quantidadeValida(produto)) {
-			Produto produtoEditado = produtoRepository.save(produto);
+		Produto produto = findById(id);
+		if (produto != null) {
+			produtoRepository.deleteById(id);
 			return true;
 		}
 		return false;
 	}
-	
-	public Produto findById(Long id){
+
+	public boolean editarProduto(Produto produto) {
+		if (quantidadeValida(produto)) {
+			produtoRepository.save(produto);
+			return true;
+		}
+		return false;
+	}
+
+	public Produto findById(Long id) {
 		Produto produto = new Produto();
-		produto = produtoRepository.findById(id).get();
+		produto = produtoRepository.findById(id).orElse(null);
 		return produto;
 	}
-	
-	
-	
-	//Validações
-	
+
+	// Validações//////////////////////////////////////////////////////
 	public boolean quantidadeValida(Produto produto) {
-		if(produto.getQuantidade() < 0) {
+		if (produto.getQuantidade() < 0) {
 			return false;
 		}
 		return true;
 	}
-	
+
+	public Integer buscarQuantidade(Long id) {
+		return produtoRepository.buscarQuantidade(id);
+
+	}
+
 }
